Remove built-in help/error handling in CmdLine
Serde.CmdLine may eventually grow this functionality, but it
looks like this is putting the cart before the horse. After the
functionality as been implemented and working we can grow the
API.
